moved logic in CardImgContainer to showImgParams function
