+++ ui/components/show/showCards/ImgContainer/CardImgContainer.module.scss
@@ -15,7 +15,8 @@
     --xxl-height: #{px2em(FindImgBoxHeight($xxl-width),'rem')};
     
     max-height: var(--max-height); 
-    height: auto; //causes CLS as images load
+    height: auto; 
+    
     max-width: px2em($width,'rem'); // this is based on the 16:9 image ratio used to determin when an image fills the width of the card vs needs padding.
     
     @include mq('xl'){
@@ -43,4 +44,16 @@
 
 .cardImgFullWidth{
     width: 100%;
+}
+
+.cardImgParamsPending{
+    // Since we don't know how much space the image will take yet, we need to set a min-height to minimize CLS
+    min-height: var(--max-height); 
+    
+    @include mq('xl'){
+        min-height: var(--xl-height); 
+    }
+    @include mq('xxl'){
+        min-height: var(--xxl-height); 
+    }   
 }
