+++ ui/functions/showImgParams.tsx
@@ -1,13 +1,16 @@
 import { Show } from "../hooks/queries/shows"
 
-export function showImgParams(show: Show): Promise<{
+export default function showImgParams(show: Show): Promise<{
   baseUrl: string | null;
   defaultImage: string | null;
   imageSizes: Array<{ w: number; h: number }> | null;
   w2HRatio: number | null;
   needsPadding: boolean | null;
+  displaySizes: string | null;
 }> {
+
   return new Promise((resolve, reject) => {
+
     if (!show.sizes) {
       resolve({
         baseUrl: null,
@@ -15,6 +18,7 @@ export function showImgParams(show: Show): Promise<{
         imageSizes: null,
         w2HRatio: null,
         needsPadding: null,
+        displaySizes: null,
       });
     }
 
@@ -28,7 +32,36 @@ export function showImgParams(show: Show): Promise<{
 
     const needsPadding: boolean = w2HRatio < 16 / 9 ? true : false;
 
-    resolve({ baseUrl, defaultImage, imageSizes, w2HRatio, needsPadding });
+    let displaySizes = "(min-width: 85.375rem) 24.125rem, (min-width: 96rem) 24.875rem, 21rem"; //based on css values for full width of container
+    
+    // Margin images, that needsPadding will be of different widths and have a different `displaySizes` value. 
+    // Let's calculate the display width for the `sizes` attribute in the picture tag.
+    //numbers used in calculation rely in css width/height values.
+    if (needsPadding){
+        const cardBreakPointWidths= [{'bp': 85.375, 'w': 24.125}, {'bp': 96, 'w': 24.875}, {'bp': 0, 'w': 21}]
+
+        const totalVerticalPadding = 24/16; //Based on css px margin. Converted for use with rem
+        // const containerHeight = 336/1.77 // Based on css
+        // const imgDisplayHeight = containerHeight - totalVerticalPadding;
+        // displaySizes = (imgDisplayHeight * w2HRatio)/16 + 'rem';
+
+        // create srcset sizes string using each breakpoint
+        let displaySizesArr = Array<string>();
+        cardBreakPointWidths.forEach((size)=>{
+            const containerHeight = size.w/1.77
+            const imgDisplayHeight = containerHeight - totalVerticalPadding;  
+            if (size.bp !== 0){
+                displaySizesArr.push(`(min-width: ${size.bp}rem) ${imgDisplayHeight * w2HRatio}rem`);
+            } else {   
+                displaySizesArr.push(`${imgDisplayHeight * w2HRatio}rem`); 
+            }
+        });
+        displaySizes = displaySizesArr.join(', ');
+
+    }
+
+
+    resolve({ baseUrl, defaultImage, imageSizes, w2HRatio, needsPadding, displaySizes });
   });
 }
