+++ ui/pages/_app.tsx
@@ -9,16 +9,6 @@ import { HydrationProvider} from "react-hydration-provider";
 import { useScrollRestoration } from "../hooks/next-restore-scroll-position";
 import { useRouter } from "next/router";
 
-// check localStorage for userId
-// if not present, create a new one
-//  var userId = localStorage.getItem('userId');
-//   if (!userId) {
-//     userId = Math.random().toString(36).substring(7);
-//     localStorage.setItem('userId', userId);
-//   }
-
-
-
 
 
 function App({ Component, pageProps,  }: AppProps) {
